+++ Main.py
@@ -23,53 +23,51 @@ def function(option,user_args = None):
     renorm.renormalize(userInput[0],userInput[1],userInput[2],userInput[3])
 
     time.sleep(0.01)
+#Makes plot
+    try:
+        if user_args.checkplot():
+            ided.pltFileExp(option,userInput[6],userInput[4],userInput[0],userInput[1],userInput[2],userInput[7],userInput[3],UI,user_args)
+    except AttributeError: ## GUI used
+        ided.pltFileExp(option,userInput[6],userInput[4],userInput[0],userInput[1],userInput[2],userInput[7],userInput[3],UI)
 
-#Makes plot (.012 s)
-    ided.pltFileExp(option,userInput[6],userInput[4],userInput[0],userInput[1],userInput[2],userInput[7],userInput[3],UI)
-
-    #FIXME if change
-    if UI:
-    #if True:
     #This code creates the .git file which is the actual plot
+    if UI == True or user_args.checkplot():
         os.chdir("Output/gnuPlot")
         directory = os.getcwd()
+    
+        #newest = max(glob.iglob(directory+"/*.plt"),key=os.path.getctime)
+        #newest = newest.replace(os.getcwd()+"/","")
+        #os.system("gnuplot %s >> /dev/null" % newest) ## >> supresses the gnuplot PNG unicode in the terminal
+    
         try:
             newest = max(glob.iglob(directory+"/*.plt"),key=os.path.getctime)
             newest = newest.replace(os.getcwd()+"/","")
-            os.system("gnuplot "+newest)
+            os.system("gnuplot %s >> /dev/null" % newest) ## >> supresses the gnuplot PNG unicode in the terminal
         except:
             print('No new plot')
-    if UI:
-#This code puts restarts the program so it can be used again
-        os.chdir("..")
-        os.chdir("..")
-        os.system("python3 Main.py "+option)
-        newest = "Output/gnuPlot/"+newest.replace(".plt",".gif")
-        if os.path.isfile(newest):
-            os.system("rm "+newest)
+        os.chdir('../..')
+        if UI:
+    #This code puts restarts the program so it can be used again
+            os.system("python3 Main.py "+option)
+            newest = "Output/gnuPlot/"+newest.replace(".plt",".gif")
+            if os.path.isfile(newest):
+                os.system("rm "+newest)
+        else:
+            pass
         try:
             os.system("mv Output/gnuPlot/*.dat Output/gnuPlot/OutputData")
             os.system("mv Output/gnuPlot/*.plt Output/gnuPlot/OutputData")
         except:
-            pass
+            raise
     else:
-        pass
-
+        try:
+            os.system("mv Output/gnuPlot/*.dat Output/gnuPlot/OutputData")
+        except:
+            raise
 ## CENDETcmd calls Main.function(...) on its own
 ## This will stop function(gui_option) from runnning when CENDETcmd imports this file
 gui_option = sys.argv[-1]
 if gui_option in ['one','two','three']:
     function(gui_option)
 else:
-    print('nothing to see here')
-
-#if __name__ == "__Main__":
-#    float('connor licks butts')
-#    option = sys.argv[-1]
-#    function(option)
-
-
-#if UI, then option = sysarg-1
-#    option = sys.argv[-1]
-#
-#    function(option)
+    pass
